CB-9874 Make cdp public endpoint service locally debuggable
So far debugging of CB public endpoints' code was not resolved. The public endpoint services however are provisioned for the usage of debug flag and debug port.
This is now introduced in CBD.

There were wrongly configured hosts (cb and autoscale) in public services docker containers, this is also going to be fixed.